+++ db/migrate/20170426004751_add_inventory_to_stacks.rb
@@ -1,0 +1,5 @@
+class AddInventoryToStacks < ActiveRecord::Migration[5.0]
+  def change
+    add_reference :stacks, :inventory, foreign_key: true
+  end
+end